requirements.txt
# Core Framework
streamlit>=1.28.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langgraph>=0.0.20

# AI/ML Libraries
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.30.0

# API Clients
requests>=2.31.0
wikipedia>=1.4.0

# Vector Store & Embeddings
huggingface-hub>=0.16.4

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Web Framework Dependencies
altair>=5.0.0
click>=8.1.0
importlib-metadata>=6.0.0
pillow>=9.5.0
protobuf>=4.23.0
pyarrow>=12.0.0
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
rich>=13.4.0
six>=1.16.0
toml>=0.10.2
tornado>=6.3.0
tzdata>=2023.3
urllib3>=2.0.0
validators>=0.20.0

# Development Dependencies (Optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Additional Dependencies for Production
gunicorn>=21.2.0
watchdog>=3.0.0